refactor: move DateTime into it's own module, create test_utils module
